+++ code/game/objects/structures/crates_lockers/largecrate.dm
@@ -1,53 +1,50 @@
-/obj/structure/largecrate
+/obj/structure/largecrate //TODO: Refactor this into a crate subtype.
 	name = "large crate"
 	desc = "A hefty wooden crate."
-	icon = 'icons/obj/storage.dmi'
+	icon = 'icons/obj/crate.dmi'
 	icon_state = "densecrate"
-	density = 1
-	atom_flags = ATOM_FLAG_CLIMBABLE
+	matter = list(MATERIAL_WOOD = 10)
+	density = TRUE
 
-/obj/structure/largecrate/Initialize()
-	. = ..()
-	for(var/obj/I in src.loc)
-		if(I.density || I.anchored || I == src || !I.simulated)
-			continue
-		I.forceMove(src)
-
-/obj/structure/largecrate/attack_hand(mob/user as mob)
-	to_chat(user, "<span class='notice'>You need a crowbar to pry this open!</span>")
+/obj/structure/largecrate/attack_hand(mob/user)
+	to_chat(user, SPAN_NOTICE("You need a crowbar to pry this open!"))
 	return
 
-/obj/structure/largecrate/attackby(obj/item/W as obj, mob/user as mob)
-	if(isCrowbar(W))
-		new /obj/item/stack/material/wood(src)
-		var/turf/T = get_turf(src)
-		for(var/atom/movable/AM in contents)
-			if(AM.simulated) AM.forceMove(T)
-		user.visible_message("<span class='notice'>[user] pries \the [src] open.</span>", \
-							 "<span class='notice'>You pry open \the [src].</span>", \
-							 "<span class='notice'>You hear splitting wood.</span>")
-		qdel(src)
+/obj/structure/largecrate/attackby(obj/item/I, mob/user)
+	if(QUALITY_PRYING in I.tool_qualities)
+		if(I.use_tool(user, src, WORKTIME_NORMAL, QUALITY_PRYING, FAILCHANCE_EASY, required_stat = STAT_ROB))
+			drop_materials(drop_location())
+			var/turf/T = get_turf(src)
+			for(var/atom/movable/AM in contents)
+				if(AM.simulated) AM.forceMove(T)
+			user.visible_message(SPAN_NOTICE("[user] pries \the [src] open."), \
+								 SPAN_NOTICE("You pry open \the [src]."), \
+								 SPAN_NOTICE("You hear splitting wood."))
+			qdel(src)
 	else
 		return attack_hand(user)
 
 /obj/structure/largecrate/mule
 	name = "MULE crate"
 
+/*
 /obj/structure/largecrate/hoverpod
 	name = "\improper Hoverpod assembly crate"
 	desc = "It comes in a box for the fabricator's sake. Where does the wood come from? ... And why is it lighter?"
 	icon_state = "mulecrate"
 
-/obj/structure/largecrate/hoverpod/attackby(obj/item/W as obj, mob/user as mob)
-	if(isCrowbar(W))
-		var/obj/item/mecha_parts/mecha_equipment/ME
-		var/obj/mecha/working/hoverpod/H = new (loc)
+/obj/structure/largecrate/hoverpod/attackby(obj/item/I, mob/user)
+	if(QUALITY_PRYING in I.tool_qualities)
+		if(I.use_tool(user, src, WORKTIME_NEAR_INSTANT, QUALITY_WELDING, FAILCHANCE_EASY, required_stat = STAT_MEC))
+			var/obj/item/mech_equipment/ME
+			var/mob/living/exosuit/working/hoverpod/H = new (loc)
 
-		ME = new /obj/item/mecha_parts/mecha_equipment/tool/hydraulic_clamp
-		ME.attach(H)
-		ME = new /obj/item/mecha_parts/mecha_equipment/tool/passenger
-		ME.attach(H)
-	..()
+			ME = new /obj/item/mech_equipment/clamp
+			ME.attach(H)
+			ME = new /obj/item/mech_equipment/tool/passenger
+			ME.attach(H)
+		..()
+*/
 
 /obj/structure/largecrate/animal
 	icon_state = "mulecrate"
@@ -56,13 +53,8 @@
 
 /obj/structure/largecrate/animal/New()
 	..()
-	if(held_type)
-		for(var/i = 1;i<=held_count;i++)
-			new held_type(src)
-
-/obj/structure/largecrate/animal/mulebot
-	name = "Mulebot crate"
-	held_type = /mob/living/bot/mulebot
+	for(var/i = 1;i<=held_count;i++)
+		new held_type(src)
 
 /obj/structure/largecrate/animal/corgi
 	name = "corgi carrier"
@@ -87,3 +79,19 @@
 	name = "chicken crate"
 	held_count = 5
 	held_type = /mob/living/simple_animal/chick
+
+/obj/structure/largecrate/animal/giant_spider
+	name = "Senshi Spider crate"
+	held_type = /mob/living/carbon/superior_animal/giant_spider
+
+/obj/structure/largecrate/animal/nurse_spider
+	name = "Kouchiku Spider crate"
+	held_type = /mob/living/carbon/superior_animal/giant_spider/nurse
+
+/obj/structure/largecrate/animal/hunter_spider
+	name = "Sokuryou Spider crate"
+	held_type = /mob/living/carbon/superior_animal/giant_spider/hunter
+
+/obj/structure/largecrate/animal/bluespace_roach
+	name = "Unbekannt Roach crate"
+	held_type = /mob/living/carbon/superior_animal/roach/bluespace
