+++ code/modules/organs/external/stump.dm
@@ -1,27 +1,39 @@
 /obj/item/organ/external/stump
 	name = "limb stump"
-	icon_name = ""
-	dislocated = -1
+	nerve_struck = -1
 
-/obj/item/organ/external/stump/New(var/mob/living/carbon/holder, var/internal, var/obj/item/organ/external/limb)
+/obj/item/organ/external/stump/New(var/mob/living/carbon/holder, var/OD, var/obj/item/organ/external/limb)
 	if(istype(limb))
 		organ_tag = limb.organ_tag
 		body_part = limb.body_part
 		amputation_point = limb.amputation_point
 		joint = limb.joint
-		parent_organ = limb.parent_organ
-	..(holder, internal)
+		parent_organ_base = limb.parent_organ_base
+		wounds = limb.wounds.Copy()
+	..(holder, null)
 	if(istype(limb))
 		max_damage = limb.max_damage
-		if((limb.robotic >= ORGAN_ROBOT) && (!parent || (parent.robotic >= ORGAN_ROBOT)))
-			robotize() //if both limb and the parent are robotic, the stump is robotic too
+		if(BP_IS_ROBOTIC(limb) && (!parent || BP_IS_ROBOTIC(parent)))
+			nature = MODIFICATION_SILICON
+
+
+/obj/item/organ/external/stump/get_tally()
+	return 4
+
+/obj/item/organ/external/stump/get_cache_key()
+	return "Stump"
 
 /obj/item/organ/external/stump/is_stump()
-	return 1
+	return TRUE
+
+/obj/item/organ/external/stump/update_icon()
+	return
 
 /obj/item/organ/external/stump/removed()
 	..()
-	qdel(src)
+	if(owner)
+		qdel(src)
+	owner = null //To stop infinate deletion loop.
 
 /obj/item/organ/external/stump/is_usable()
-	return 0
+	return FALSE
